Fix bug where if a file uploaded from local to cloud is removed from local
Previous:

Events
1. touch foo
2. sync: local added foo. cloud does not have it => add to cloud
3. rm foo
4. sync: When compared to last scan, local removed foo and remote added foo. To play safe add foo to
  local

Now:

When uploading blob to remote, add a mac address based client id as as metadta
In the case mentioned above at step4 we will also see that foo's client id is same as the local
machine. This means foo was removed after foo was uploaded from the local machine. So the local
remove should win => remove the blob