Do a proper 2 way sync
Before this change, i was computing diff on both ends individually and trying to merge just the
diffs. This works as long as you dont update "previous state" upon seeing any failure. I was
mostly doing that, but i feel there will be more cases where "i start an action, then realize
midway the action should not be done and this can be the happy path". So it makes sense to
be prepared for that => i should try to make remote and local the same - as part of doing that
i will use informaiton about local diff, remote diff from last time.